(MOT-4107) refactor(harness): split integration scenario compiler modules
