+++ harness/evals/integration/src/expand/render.rs
@@ -1,0 +1,43 @@
+use anyhow::Context;
+use serde_json::json;
+
+use super::{CompiledFixtureV1, Placeholders};
+use crate::types::scenario::IntegrationScenarioV1;
+
+pub(super) fn validate_placeholders(fixture: &CompiledFixtureV1) -> anyhow::Result<()> {
+    render_compiled(fixture)
+        .map(|_| ())
+        .context("validating compiled placeholders")
+}
+
+/// Deterministic, fully expanded representation used by the `render` CLI.
+pub fn render_compiled(fixture: &CompiledFixtureV1) -> anyhow::Result<String> {
+    let run_id = format!("render-{}", fixture.scenario.id.to_ascii_lowercase());
+    let session_id = format!(
+        "render-session-{}",
+        fixture.scenario.id.to_ascii_lowercase()
+    );
+    let base = Placeholders::new(&run_id, &session_id);
+    let system_prompt = base.expand_str(&fixture.system_prompt_template)?;
+    let digest = crate::canonical::sha256_of_bytes(system_prompt.as_bytes());
+    let placeholders = base.with_system_prompt_sha256(&digest);
+
+    let mut scenario = serde_json::to_value(&fixture.scenario)?;
+    placeholders.expand_value(&mut scenario)?;
+    let mut script = serde_json::to_value(&fixture.script)?;
+    placeholders.expand_value(&mut script)?;
+    Ok(crate::canonical::canonical_json_pretty(&json!({
+        "scenario": scenario,
+        "router_script": script,
+        "system_prompt": system_prompt
+    })))
+}
+
+/// Serialize an authored scenario in stable YAML for `init`.
+pub fn render_authored_yaml(scenario: &IntegrationScenarioV1) -> anyhow::Result<String> {
+    let mut rendered = serde_yaml::to_string(scenario)?;
+    if !rendered.ends_with('\n') {
+        rendered.push('\n');
+    }
+    Ok(rendered)
+}
